Fixed autoincrement for the SQL database

+++ uqcsbot/models.py
@@ -20,7 +20,7 @@ class AOCWinners(Base):
     __tablename__ = "aoc_winners"
 
     id: Mapped[int] = mapped_column(
-        "id", BigInteger, primary_key=True, nullable=False, autoincrement=True
+        "id", Integer, primary_key=True, nullable=False, autoincrement=True
     )
     aoc_userid: Mapped[int] = mapped_column("aoc_userid", Integer, nullable=False)
     year: Mapped[int] = mapped_column("year", Integer, nullable=False)
@@ -31,7 +31,7 @@ class AOCRegistrations(Base):
     __tablename__ = "aoc_registrations"
 
     id: Mapped[int] = mapped_column(
-        "id", BigInteger, primary_key=True, nullable=False, autoincrement=True
+        "id", Integer, primary_key=True, nullable=False, autoincrement=True
     )
     aoc_userid: Mapped[int] = mapped_column("aoc_userid", Integer, nullable=False)
     year: Mapped[int] = mapped_column("year", Integer, nullable=False)
